python/requirements.txt
Pillow>=10.0.0
matplotlib>=3.7.0
numpy>=1.24.0
imageio>=2.31.0
imageio-ffmpeg>=0.4.8
numba>=0.58.0
psutil>=5.9.0
# PyTorch 2.9.1 + CUDA: run install_requirements.ps1 for GPU (RTX 5060)
